renamed tablerendering, modified table.go
