+++ render/render_test.go
@@ -1,4 +1,4 @@
-package tablerendering
+package render
 
 import (
 	"bytes"
@@ -8,10 +8,10 @@ import (
 
 func TestWithByteBuffer(t *testing.T) {
 	out := new(bytes.Buffer)
-	RenderTable(out, []string{"test", "version", "text"}, [][]string{{"1", "version 1", "empty"}, {"a2", "b2", "c3"}})
+	Table(out, []string{"test", "version", "text"}, [][]string{{"1", "version 1", "empty"}, {"a2", "b2", "c3"}})
 
 	countedLines := strings.Count(out.String(), "\n")
-	if countedLines != 4 {
+	if countedLines != 3 {
 		t.Fail()
 	}
